refactor: Add RefreshInterceptor to handle token refreshing
This commit adds the RefreshInterceptor to the application. The RefreshInterceptor is responsible for intercepting HTTP requests and checking if the token needs to be refreshed. If the token is expired, the interceptor will automatically refresh it before proceeding with the request. This enhancement improves the security and reliability of the application by ensuring that the token is always up to date.