refactor: Add karmaConfig to angular.json and package.json
This commit adds the "karmaConfig" property to the "angular.json" file and the "test" script in the "package.json" file. The "karmaConfig" property is set to "karma.conf.js" in the "angular.json" file, and the "test" script is updated to include the "--karma-config karma.conf.js" flag in the "package.json" file. These changes ensure that the Karma configuration file is properly referenced and used during testing.